migrate core tests to integration tests
